feat: add init db

+++ src/models/ratings.ts
@@ -11,7 +11,7 @@ export interface RatingsEntity extends BaseEntity {
 export const RatingsModel = new BaseModel<RatingsEntity>('ratings', (table) => {
   table.integer('user_id').references('id').inTable('users').notNullable();
   table.integer('movie_id').references('id').inTable('movies').notNullable();
-  table.integer('rating').notNullable();
+  table.float('rating').notNullable();
   table.unique(['user_id', 'movie_id']);
   table.timestamp('timestamp').defaultTo(knex.fn.now());
 });
